+++ backend/substrapp/tasks/tasks_compute_plan.py
@@ -57,8 +57,16 @@ def delete_cp_pod_and_dirs_and_optionally_images(channel_name, compute_plan):
             compute_plan_key=compute_plan_key,
         )
         test_tasks = client.query_tasks(category=computetask_pb2.TASK_TEST, compute_plan_key=compute_plan_key)
-    algo_keys = [x["key"] for x in algos]
-    metric_keys = [key for task in test_tasks for key in task["test"]["metric_keys"]]
+    algo_keys = [Algo(channel_name, x) for x in algos]
+
+    # TODO: make it possible in the orchestrator to retrieve metrics used in a compute plan similar as above
+    # with other algo kinds
+    metrics: dict[str, Algo] = {}
+    with get_orchestrator_client(channel_name) as client:
+        for task in test_tasks:
+            for key in task["test"]["metric_keys"]:
+                if key not in metrics.keys():
+                    metrics[key] = Algo(channel_name, client.query_algo(key))
 
     # See lock function PyDoc for explanation as to why this lock is necessary
     with get_compute_plan_lock(compute_plan_key):
@@ -76,10 +84,9 @@ def delete_cp_pod_and_dirs_and_optionally_images(channel_name, compute_plan):
         teardown_compute_plan_dir(dirs)
 
     _remove_docker_images(algo_keys)
-    _remove_docker_images(metric_keys)
+    _remove_docker_images(list(metrics.values()))
 
 
-def _remove_docker_images(keys: list[str]):
-    for key in keys:
-        image_tag = Algo.image_tag(key)
-        delete_container_image_safe(image_tag)
+def _remove_docker_images(algos: list[Algo]):
+    for algo in algos:
+        delete_container_image_safe(algo.container_image_tag)