+++ backend/substrapp/tests/compute_tasks/test_algo.py
@@ -4,4 +4,4 @@ from substrapp.compute_tasks.algo import Algo
 def test_algo_container_tag(algo: Algo):
     tag = algo.container_image_tag
 
-    assert tag == "algo-ae0fa090"
+    assert tag == "algo-f2ca1bb6c7e907d0"