+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - (BREAKING) Store enum values as string (impact ordering) - Algo (#1040)
 - (BREAKING) Add mandatory name field to compute plan (#1027)
+- Use the algo checksum as image tag (#1019)
 
 ### Fixed
 - completed pod state in `kubernetes_utils.watch_pod()` (#1046)